+++ typescript/packages/wallets/crossmint/src/chains.ts
@@ -23,6 +23,8 @@ import {
     shapeSepolia,
     skaleNebula,
     skaleNebulaTestnet,
+    story,
+    storyTestnet,
     victionTestnet,
     xai,
     xaiTestnet,
@@ -81,6 +83,8 @@ export const mintingChains = [
     "shape-sepolia",
     "skale-nebula",
     "skale-nebula-testnet",
+    "story-mainnet",
+    "story-testnet",
     "soneium-minato-testnet",
     "xai",
     "xai-sepolia-testnet",
@@ -90,6 +94,10 @@ export const mintingChains = [
 ] as const;
 
 export type SupportedMintingChains = (typeof mintingChains)[number];
+
+export function isStoryChain(chain: string): boolean {
+    return chain === "story-mainnet" || chain === "story-testnet";
+}
 
 const chainMap: Record<SupportedFaucetChains | SupportedSmartWalletChains | SupportedMintingChains, Chain> = {
     arbitrum,
@@ -113,6 +121,8 @@ const chainMap: Record<SupportedFaucetChains | SupportedSmartWalletChains | Supp
     "shape-sepolia": shapeSepolia,
     "skale-nebula": skaleNebula,
     "skale-nebula-testnet": skaleNebulaTestnet,
+    "story-mainnet": story,
+    "story-testnet": storyTestnet,
     "viction-testnet": victionTestnet,
     xai,
     "xai-sepolia-testnet": xaiTestnet,