+++ typescript/packages/plugins/avnu/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @goat-sdk/plugin-avnu
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+  - @goat-sdk/wallet-starknet@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes