+++ typescript/packages/plugins/cosmosbank/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @goat-sdk/plugin-cosmosbank
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+  - @goat-sdk/wallet-cosmos@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes