+++ typescript/packages/plugins/debridge/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @goat-sdk/plugin-debridge
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+  - @goat-sdk/wallet-evm@0.2.10
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [9371f33]
+  - @goat-sdk/wallet-evm@0.2.9
+
 ## 0.0.4
 
 ### Patch Changes