+++ typescript/packages/plugins/dexscreener/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @goat-sdk/plugin-dexscreener
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+
 ## 0.1.5
 
 ### Patch Changes