+++ typescript/packages/plugins/plunderswap/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @goat-sdk/plugin-plunderswap
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+  - @goat-sdk/wallet-zilliqa@0.3.3
+
+## 1.1.2
+
+### Patch Changes
+
+- @goat-sdk/wallet-zilliqa@0.3.2
+
+## 1.1.1
+
+### Patch Changes
+
+- @goat-sdk/wallet-zilliqa@0.3.1
+
 ## 1.1.0
 
 ### Minor Changes