+++ typescript/packages/plugins/zilliqa/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @goat-sdk/plugin-zilliqa
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+  - @goat-sdk/wallet-evm@0.2.10
+  - @goat-sdk/wallet-zilliqa@0.3.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [9371f33]
+  - @goat-sdk/wallet-evm@0.2.9
+  - @goat-sdk/wallet-zilliqa@0.3.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @goat-sdk/wallet-zilliqa@0.3.1
+
 ## 1.0.0
 
 ### Patch Changes