+++ typescript/packages/wallets/fuel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @goat-sdk/wallet-fuel
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [3091a49]
+  - @goat-sdk/core@0.4.8
+
 ## 0.1.7
 
 ### Patch Changes