+++ typescript/packages/wallets/viem/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @goat-sdk/wallet-viem
 
+## 0.2.11
+
+### Patch Changes
+
+- @goat-sdk/wallet-evm@0.2.10
+
+## 0.2.10
+
+### Patch Changes
+
+- 9371f33: Extract ENS resolution to its own package
+- Updated dependencies [9371f33]
+  - @goat-sdk/wallet-evm@0.2.9
+
+## 0.2.9
+
+### Patch Changes
+
+- d7ecfe6: Upgrade viem and add story protocol
+
 ## 0.2.8
 
 ### Patch Changes